+++ vector/src/main/java/im/vector/riotx/features/home/room/detail/timeline/factory/EncryptionItemFactory.kt
@@ -65,7 +65,7 @@ class EncryptionItemFactory @Inject constructor(private val stringProvider: Stri
 
     private fun buildNoticeText(event: Event, senderName: String?): CharSequence? {
         return when {
-            EventType.ENCRYPTION == event.getClearType() -> {
+            EventType.STATE_ROOM_ENCRYPTION == event.getClearType() -> {
                 val content = event.content.toModel<EncryptionEventContent>() ?: return null
                 stringProvider.getString(R.string.notice_end_to_end, senderName, content.algorithm)
             }