+++ CHANGES.md
@@ -13,6 +13,7 @@ Improvements 🙌:
  - Add settings for rageshake sensibility
  - Fix autocompletion issues and add support for rooms, groups, and emoji (#780)
  - Show skip to bottom FAB while scrolling down (#752)
+ - Enable encryption on a room, SDK part (#212)
 
 Other changes:
  - Change the way RiotX identifies a session to allow the SDK to support several sessions with the same user (#800)